// DesignParkinglot/code/src/enum.ts
export enum VehicleType {
  BIKE = "BIKE",
  CAR = "CAR",
  BUS = "BUS",
}

export enum SpotSize {
  SMALL = "SMALL",
  MEDIUM = "MEDIUM",
  LARGE = "LARGE",
}

export enum TicketStatus {
  ACTIVE = "ACTIVE",
  PAID = "PAID",
  EXPIRED = "EXPIRED",
}

export enum SpotStatus {
  AVAILABLE = "AVAILABLE",
  OCCUPIED = "OCCUPIED",
  OUT_OF_ORDER = "OUT_OF_ORDER",
}

export enum PaymentMethod {
  CASH = "CASH",
  UPI = "UPI",
  CARD = "CARD",
}
export enum PaymentStatus {
  SUCCESS = "SUCCESS",
  FAILED = "FAILED",
}